fix zeros
Previous versions assumed that all missing values for a reach were zeros. That was incorrect. This version corrects that by using the information in fsh_perPass$targetTaxaPresent. Now any NA's in a pass are either coded as a true zero (if targetTaxaPresent == "N") or they are coded as an NA and that sample is removed. This removed about 5% of samples.